Use Velum image from registry.suse.de
This is not an ideal change, as we no longer separate the image pull from the
rebuild. Given how slow the Provo connection to Nue can be, this may become
a problem.

For now, lets give it a shot so the Build Service team can revert the temporary
change they have applied for us.